Refactor app.py for improved functionality and clarity

+++ app.py
@@ -18,7 +18,7 @@ def parse_pos(key):
     if isinstance(key,list):
         return (key[0],key[1])
     x, z = key.strip("[]").split(",")
-    return int(x), int(z)
+    return float(x), float(z)
 
 
 def is_valid_tile(world, pos):
